[FIX] m.VariantManagement: fix for save in 'Manage Dialog'
The issue here is, that it is possible to leave the Manage Views dialog
even when the dialog contains errors like empty view titles or view
titles which contains not allowed characters.
This may even lead to exceptions during the re-appliance of the views
based on the usage of non-allowed characters in its titles.

This change prohibits the save when such a situation as explained above
occurs.

Change-Id: I993d73c7931ef47e482756208ccd8d3609136ec4
SNOW: CS20230006763764
CR-Id: 002075125900000124102024