[FIX] sap/ui/core/Theming: Ensure change event is fired after changes
For use cases where the Theming.change event is processed synchronously
and another Theming.change event is triggered within the event handler
of the first change event not all change events are processed corretly.
Remove collecting the change events since this is not necessary anymore
(was needed previously for setting theme and themeroot together).

SNOW: DINC0031219
Change-Id: I6acb857466dadd06c0992220ea1373920c6abc38