[INTERNAL] sap.ui.core: avoid deletion of shared DebugMode.qunit script
The DebugMode.qunit.js script is used by two test pages, by the async DebugModeAsync.qunit.html page and by the sync DebugMode.qunit.html
page (which is legacy). As the script has the same name as the legacy,
sync page, it is deleted during the 2.0 transformation, too.

To avoid this, the sync page is renamed to DebugModeSync.qunit.html.

Change-Id: Ibc3dc3899a54238ce8c81c3fc743ec6b45d46813