[INTERNAL] sap.ui.test: Configuration of unstable test adjusted
Problem:
Unstable test for iFrameLogging because iStartMyAppInAFrame timeouts
(the iframe could not be loaded within the default timeout of 80ms)
The problem cannnot be reproduced when running the test in the
local browser, but only in the infrastructure, so cannot be inspected
in more detail.

Solution:
Since autowait is disabled by default for iStartMyAppInAFrame, enable
autowait as first possible solution. Also add minor improvement in the
fallback check for framework initializaton for the older UI5 versions

SNOW: DINC0035964
Change-Id: Ib21e5b3c1c9d2a9e82586c6c2fd844fbe5a17861