[INTERNAL] CellSelector: Optimize _selectCells calls
Previously, _selectCells was called in places where it did not need to be called.

A) Scrolling: When scrolling or the control updates, _selectCells has been called to ensure that the selection is correctly drawn when e.g. scrolling. Instead of calling _selectCells, we now just call _drawSelection
B) Hovering the same cell as before: There was no early exit, while selecting and hovering the same cell as before. This lead to a _selectCells call although it is not required as the cell did not change. Therefore, there is a check now to ensure _selectCells is only called when necessary.

Change-Id: I6e954ed8a2ff4ef1bf6de68ba45c8bb611b6fb69