[FIX] sap.f.FlexibleColumnLayout: Rounding issue corected
Problem:
When the FCL converts the column percent widths from floats
to integers, sometimes the total width does not equal 100.
This is because the converion to integer was made using parseInt,
which always floors the parsed values

Solution:
Replace parseInt with parseFloat + Math.round. In addition,
add a normalization function for the cases where Math.round
is not sufficent to ensure total width of 100 (the qunit
reproduces such a case)

Change-Id: I16af25012a2ab2c7004d38912f014c8c656b4fff