[INTERNAL][FIX] sap.ui.dt: fix z-index calculation
Overlay z-index calculation used a jQuery method that doesn't only take the z-index from the computed styles of the element but goes further up the DOM tree until it finds an element with a z-index or until the root element is reached (defaulting to z-index = 0)
This is problematic if stacking context boundaries are trespassed, as the z-index would no longer be relative to the current stacking context.
As a result some overlays are not reachable, since the overlay structure doesn't fully represent the element structure.
This fix changes the behavior to only go up until the next UI5 element is reached, since it already represents its own z-index, thus going any further makes no sense.

Change-Id: I13710d5c7d7db6c3bd8fd207ac4ec39bcb044d36
CR-Id: 002075125900000434842024
SNOW: CS20230006470167