[INTERNAL][FIX] FileSizeFormat: fix call to ResourceBundle.getText
ResourceBundle.getText expects the placeholder values as an array.
Providing an individual value might result in wrong formatting. E.g. when
a falsy value is provided, getText might interpret this as
'no placeholder values' and the whole replacement step is skipped.

Note: This was revealed by the linked incident, but this change does not fix it.

PS1: Fix
PS2: Add test

SNOW: DINC0039665
Change-Id: I45d936e7e11f8bd7bb56e310c8352a3808550256