[INTERNAL] Operator: allow symbol as text
Situation:
If a Text starting with "=" is entered in a Field, the "=" is removed. A filtering for text starting with "=" on the typeahead is not possible.
In FilterField it is not possible to enter a not validated equal token. If "=A" is entered and there is a match in the typeahead, let's say "ABC", the autocomplete functionality replaes the entered text and removes the "=". If there is no match in typeahead an error is shown instead of creating a not validated EQ-condition.
It is not possible to enter text with multiple "=", like "===A".

Changes:
1. The "hideOperator"-mode will also used in parsing of an Operator.
	That means, in Field case "=" is not used as operator symbol but just as entered text. So typeahead also uses "=" to search. The "=" will not be removed from entered text.

2. If operator ("=") entered in FilterField it will be used as Operator.
	Autocomplete will be diabled in this case.
	If there is a exact match for the text after the "=" a validated condition will be created. Otherwise a not-validated condition with the entered text will be created.

3. EQ-Operator now only checks for the first "=". So entering "==" is valid now.

Change-Id: I1412742850608ccfe68a02a1e2f84b7e6ca0a600