[INTERNAL] sap/ui/fl: Introducing a BTP Service Connector
In BTP we have two backends, the PERSO service and the KEYUSER service.
Now the KEYUSER service can also handle USER layer changes. With this
improvement it is now possible to reduce a backend request in BTP.
Therefor this new BTP Service Connector is needed.
Within gerrit change 6072157 and 6086524

Change-Id: I4f2a3e3835ab1bf8906f066d382e7c8ebd2ad920
JIRA: CLOUDIOI-1067
CR-Id: 002075125900000497842024
BCP: n/a