[FIX] sap.uxap.ObjectPageLayout: Fix collapsing header
Problem: On FireFox, when you scroll to the last SubSection and try to
expand the Header, in some cases it immediately gets collapsed.

Solution: The problem seems to happen due to some rounding differences
in FireFox. I am reproducing the issue, when I navigate to the last
SUbSection and after that scroll a little bit down (1px) with the mouse.
In this case, when expanding the Header, there are two scroll events
fired. Upon the second one, as the scroll hanlder is not suspended, the
Header gets collapsed.

SNOW: DINC0012992
Change-Id: Ic0150abb259dc0ff643e678e7650ec323e46a315
CR-Id: 002075125900000812532024
(cherry picked from commit 5862076fc6add34c46b2fcf45163ef67d3b721bb)