[FEATURE] sap/base/i18n/ResourceBundle: Enable caching of .properties files
.properties file requests are cached to avoid duplicate
requests for the same file.

The only case when 2 requests for the same file are performed is:
When the first attempt to load is performed async and is not yet completed
and meanwhile a sync attempt to load is done;
then because the sync needs to return the properties file, a sync request
is performed while the async request is still in progress.

In case a .properties file is not found, the request is not cached,
so every new request for the same file will be performed again.

JIRA: CPOUI5FOUNDATION-167
Change-Id: I3cc48871934bdf926acdb2506a5b68052b930d7d