[FIX] sap.m.ProgressIndicator: Fix overlapping in table
Problem: displayValue of ProgressIndicator overlaps with the progress
bar when placed in sap.ui.table.Table on horizon themes dark, hcb and
hcw.

Solution: It seems the problem has been fixed in the past, but only for
default horizon theme. In Compact mode paddings are adjusted, so that
there is no overlapping. In Condensed mode, there is not enough height
for showing the display value, so that it gets hidden.

SNOW: DINC0041428
Change-Id: I028cfce37a26a2c80f3ed54fb2d88b1bada5cca7