[INTERNAL] qunitPause: avoid race condition when accessed via qunit
The previously used `sap.ui.define` defines the module asynchronously.
In contexts, where the existence of the module is checked synchronously,
like during the startup of qunit/qunit-2, this might cause a double
execution of qunitPause which then might cause further issues.

By defining the module synchronously with a non-public API of the
loader, such issues can be prevented. As this sync API does not have to
resolve dependencies, it doesn't cause synchronous loading or code
execution.

However, the API remains strictly internal and must not be used
elsewhere as it is not supported by UI5 or AMD-based tooling.
For a pure test scenario like qunitPause, its usage seems acceptable.

SNOW: DINC0055761
Change-Id: I25da62d1f87ee93bedd5b58bfe8369bab63e5994