[INTERNAL] sap.ui.test: Update test configuration
Problem:
Iframe content cannot be loaded when the content is a
qunit test that also loads qunitPause.js and Core.js.
Loading fails as the Core.ready promise never resolves.

The root cause is unknown so far (only happens on the
test voter and never when running the test locally)

The issue is only observed when using the older qunit
version (qunit1)

Solution:
Given that this is not an expected use-case in practice
(loading a qunit test inside an iframe was only needed
as an OPA self-test), this change proposes a workaround
by updating the qunit version to the newer qunit2.

Change-Id: I8a423c17a38d1aeb96ab893052b23fdcb8d41871
SNOW: DINC0035964