[INTERNAL] CopyProvider: Enabled state of the Copy button
 - If there is no selection in the table then the copy button should be
disabled according to UX guideline.
 - CopyProvider registers selectionChange event of tables to determine
when the copy button state should change. Besides that, a binding change
might also cause a selection change which might not trigger the
selectionChange event e.g. when the ui.Table rebinds while there is
selection on the table or when a selected item is filtered out in the
m.Table case.
 - Since the CopyProvider works together with the CellSelector plugin,
we also need a selectionChange event from the CellSelector. This change
introduces a private selectionChange event for the CellSelector. This
can also be useful for the CellSelector consumers in the future.
 - For the sap.m.Table the selection can be changed via API calls for
which does not trigger selectionChange event. To handle API relevant
selection changes the ListBase control fires private itemSelectedChange
event when the selected property of a ListItem is changed.

Jira: CPOUIFTEAMB-2273

Change-Id: I2208d371192a1aab025ae1421254ad8fde3cb5ca