[INTERNAL] m.ListBase: let selection APIs to fire selectionChange event
 - In UI5 it is a general convention that API calls do not fire events,
for example, changing the value property of the sap.m.Input does not
fire the liveChange event.
 - Usually the consumer who calls the API also owns the related event
handler this means the consumer can easily call the event handler after
the API call or the #fireEvent API anyway public for every control but;
  1- The event handlers does not always belong to the one who calls the
API, for example, a generic selection change handler might need to be
triggered when the selection needs to be changed via API calls
  2- The selectionChange event has complex parameters where just calling
the #fireSelectionChange might not be suitable for the event handlers
that listen those event parameters.
 - ui.Table fires selectionChange event when selection APIs are called
here to be complatible with older version and ui.Table control we
provide an option to fire the selectionChange event with selection APIs.

Change-Id: I6bd27817c2ce637630eda7b5e48c290a1f81f2c1