[INTERNAL] m.ListBase: select all with grouping and top-level context
 - Although there is no binding context is set for GroupHeaderListItems
the context might get propagated from a top level which allows selection
API to be called on those group headers.
 - In the V4 binding context it seems like it is not allowed to call the
selection API for the top level context and this throws error.
 - Since group headers are anyway not selectable we do not only trust
that they do not have binding context but check the
ListItemBase#isSelectable API before we call the selection APIs on the
binding context of items.

SNOW: CS20240006904619

Change-Id: I0135c9d2a7199f7e6e20d5eca5e8d8082dcc39c3