[INTERNAL] sap.ui.test: Update test configuration
Problem:
Iframe content cannot be loaded when the content is a
qunit test that also loads qunitPause.js and Core.js.

The root cause is unknown so far (only happens on the
test voter and never when running the test locally)

A similar issue was observed for SNOW: DINC0035964 and
since its fix I8a423c17a38d1aeb96ab893052b23fdcb8d41871
the issue does not reproduce on the tests voter.

Solution:
Given that this is not an expected use-case in practice
(loading a qunit test inside an iframe was only needed
as an OPA self-test), this change applies the same solution
as for SNOW: DINC0035964 i.e. loading into the iframe a
test file that uses the newer qunit2 version.

SNOW: DINC0040048
Change-Id: If2e34a7181fcca97131b2a6b65eed74733c7c526