[FIX] ColumnMenu: Check for container instance when setting visibility
When changing the visibility of an item, the item visbility is updated in the menu as well.
It checks the items container, to adjust the visibility accordingly.

There is the case, that the visibility is set before the container has been initialized (dynamic binding in the view, initial value set to false, etc.).
In this case setting the visibility would lead to an exception.

SNOW: CS20240006886346
Change-Id: I795fd74dcc61b635995e8516d6cba5ded5378fe4
(cherry picked from commit da26c3bdfbf1fc3ca422001b060bce4081209202)
CR-Id: 002075125800003004732024