[FEATURE] sap.uxap.AnchorBar: Separator added
By the new design spec of IconTabBar, when there are 2 click areas for a Tab (the Main Action and the Chevron for the Menus), there has been a separator been introduced (styled like the Split Button) to visually separate the two click areas for Usability and ACC reasons.
We implemented it for AnchorBar also.

Jira: BGSOFUIPIRIN-6430
Change-Id: I8ca084def576ee607eb3e66b31e170d4a4eeae08