[FIX] sap.m.ViewSettingsDialog: 'View Settings' title is not visible on entering the filter details page
Issue: If there is only single Filters tab in the VSD, on switching
to Filter details page, 'View Settings' dialod title appears for a
while and then it is replaced by 'Filter By > ...' title.

Solution: In that case there is no need to set dialog title to
'View Settings' as it is never shown.

SNOW: DINC0051893
Change-Id: I1bef90f589fc5544006979167b12485ffb6f92e5