[INTERNAL]sap.ui.rta: Fix RuntimeAuthoring Test with invalid version
A RuntimeAuthoring test was unstable because it activated versioning
while using a test connector without versioning, triggering an exception
on the object storage when getting the version chain, as no "version"
FlexObject is saved on the storage but the application expected it to be
found.

The affected test was moved to the RuntimeAuthoring-versioning module,
where a connector with versioning is used.

Change-Id: I484242c55c474d0cbd45ac408bae8416a4cdd355