[INTERNAL] Revert of  isTrustedSource property change on the PDFViewer
New property isTrustedSource to be reverted for the PDFViewer, now the PDFViewer will display the PDF's by default.

As there are multiple teams which have not yet implemented the introduction of this property, After discussing with our security experts we have decided to revert these changes and implement the same once all applications adhere to the property change.

SNOW: DINC0059782
CR-Id: 002075125900000925232024
Change-Id: Iea0acbbbfb9dfa9798ba11e166f9c2812c0a78a8