[FIX] DemoKit: Identification of relative SEO-optimized URLs corrected
Problem:
The function that checks that a URL is in SEO-optimized format
fails to identify SEO-optimized URL if we pass to it a
relative URL without the leading "/".

Solution:
SEO-optimized URLs without a leading "/" are a legitimate
case, so corrected by defining the starting "/" as optional
in the above check.

SNOW: DINC0052966
Change-Id: I5dde8564e820140a8ceb821e9742f30abf843056