[FIX] DemoKit: Scroll to module function corrected
Problem:
In the ApiRef, scrolling to a method fails when
the method is static and belongs to a module.

The scrolling fails because the subsection that contains
the method is not found. The subsection is not found,
because we search for it by title, expecting the title to
match the method name.
However, the subsection title does not exactly match the
method name for static methods prefixed with
"module:", as we strip the "module:" prefix from the
subsection title.

Solution:
Ensure we find the subsection by searching for its actual
title.
Note: an alternative solution of using a different search
criteria was also researched but ruled-out as the current
solution appears the simplest for now and requiring least
refactoring.

Change-Id: I6010fdbbda89d565a86b344680f13d92daf5dd60