[FIX] sap.m.Avatar: Wide initials fix
Problem: In some cases when the container, which holds the initials of
the sap.m.Avatar, is with the same width as the Avatar itself, visually
it looks as the initials overflow the avatar.

Solution: Show the fallback icon, not only when the initials overflow,
but when they are with the same width as the Avatar as well.

SNOW: DINC0058161
Change-Id: I62058bd431ca6e269344d60f47f938ac36a2bf8b