[FIX] sap.m.IconTabBar: Focus of IconTabFilter is not changing
Problem: When the focus is set on IconTabFilter and then it goes out of
 the viewport, at the same time a property is set to the IconTabFilter,this causes a rerender of IconTabHeader, focus is lost, and after rendering focus is regained and the page is scrolled at the position of the IconTabFilter. When happens in a list, scrolling becomes impossible.

Solution: Added tabindex -1 to IconTabFilter. This way the IconTabHeader will not be rerender and only the property will be updated. Tabindex was added in afterrendering function and this caused a rerender because there was a difference in the attributes

SNOW: CS20230006325547
Change-Id: I7daed2ad0b706b02440b76cb4228082f946a12a0