[INTERNAL] sap.f.FlexibleColumnLayout: Store begin column width
Problem: When going from one layout to another we either use the already saved user values for the particular layout (if already visited), or we use the predefined values (as per specification). However, there are some cases, when this behavior is not very pleasant from user perspective. For example, when on ThreeColumnsMidExpanded layout user minimizes the first column to the min-width possible, then he maximizes the third column, so that the layout is switched to ThreeColumnsEndExpanded --> the first columns gets bigger than the min-width value, specified by the user in the PREVIOUS layout. Another case is similar - when switching from TwoColumnsLayout to ThreeColumnsMidExpanded layout.

Solution: Storing the begin column width value (set by the user through
dragging the separators) and use it when going from one layout to
another.

JIRA: BGSOFUIPIRIN-6417
Change-Id: Id8c666b19e09685459eea09ec996d46137f7b557