[INTERNAL][FIX]sap.ui.rta: Improve VariantManagementState Performance
Fixes a performance regression when checking the referenced variant chain while building the variant management state. Currently the check was executed in a loop for each control change, however since the referenced variants are independent of the control change this is unnecessary and can be checked outside the loop which improves performance of the check from O(n^2) to O(n).

Change-Id: Icdba23a028288824231b906306286b2d8b87d26d
CR-Id: 002075125900001162942024
SNOW: CS20240007013795