[FIX] Integration Cards: Improve card in dialog with defined width
Solving the following problem:
If ShowCard action is used with predefined "width",
the opened card was fixed to that width and when opened
on smaller screen, a scrollbar appeared.
Better behavior is to set the width as contentWidth of the Dialog,
then on smaller screen - the dialog and the card are not exceeding
the screen size.

SNOW: DINC0056482
Change-Id: I140cbfcf7ef4bf20557f80590f778d1620094ed9