[INTERNAL] sap.ui.fl handle right version when last draft change is delete
The condense route can send a single delete request to the backend. It
can be the case that this change is the last draft change. It this is
the case the backend will also delete the version file.
Need to enhance the objectStorageConnector to simulate the backend and
also need to get the correct version model from the real backend

Change-Id: I16f5ca655d27c3be1ba697bd45f7fd799a8e0188
SNOW: DINC0055428