[INTERNAL] PDFViewer now displays a message using sap.m.IllustratedMessage when the property isTrustedSource is set to 'false'
If isTrustedSource is set to true, the PDFViewer opens with the displayType set to "Embedded" on desktop devices, which means that the PDF content is directly shown within the PDFViewer. Set this property to true only when the PDF is generated by the application or the PDF is scanned for viruses.

If isTrustedSource is set to false, the PDFViewer opens with the displayType set to "Link" on desktop devices, which overrides any configuration that has been provided by the application for the property displayType. This means that the PDFViewer appears as a toolbar with a download button.

PDF's will continue to be displayed by default if the property is not set. However, in future releases, we plan to change this default setting. If not configured manually, PDFViewer will open with displayType set to "Link"

SNOW: DINC0114788
CR-Id: 002075125800001149162024
Change-Id: I49e954bbcc69ab7516722116f9785589d658acac
(cherry picked from commit 5031852378f0eeaf2bd62990e150252596d466cb)