[FIX] sap.m.Toolbar: Fix alignment in Tables
Problem: Left alignment of Toolbar in header of Tables does not correspond
to the VD spec. According to the spec for fiori_3 and horizon themes -
Toolbar in Tables' header should have 1 rem left padding. Refer to Table
Header / Title Toolbar in VD Specs.

Solution: It seems, for historical reasons, we sticked to the
Transperant Toolbar design to override the left padding in the Tables'
header. However, according to the VD spec, Transparent Toolbar should
have equal left and right paddings (0.5 rem). This was fixed 3 months
ago (as left padding was 1 rem). Fixing this, it seems we have broken
the alignment of Toobar in header of Tables. That's why, now we use the
sapMListHdrTBar class to adjust the left padding of the Toolbar, placed
the header of the Tables.

SNOW: DINC0066722
Change-Id: I9311e4582c5285e4a4c6df6e8518297568be1b8b