[FEATURE] sap.m.PlanningCalendar: render disabled dates in Week & OneMonth views
Issue: When there is a planning calendar with a MinDate set, Weeks and
OneMonth views would start at the minimum date not the beggining of the
week/month as it would otherwise.

Solution: Change the start date for those views to the start of the
week/month and show dates as disabled.

JIRA: BGSOFUIBALKAN-7722
Change-Id: I0011d2ecdd1a0c3429c5c9066312e5b60d998575