[FIX]sap.m.ComboBox: Seleted item is scrolled into view
When a suggestion item placed very low on the list of suggestions is selected
it was not visible after reopening the suggestions list - it was not
scrolled into view. This was caused because calling the applyPosition of the popover
during ValueStateHeader onAfterRendering can not correctly scroll the popover content
to the desired location.

SNOW: DINC0069440
Change-Id: Ib801c8118282f2201cdbde9ee23bbe60607cea09