[INTERNAL][FIX] sap.ui.layout: Fix display indicator for HideSimpleForm
HideSimpleForm changes don't contain any info about the group where the
field was removed from it can only be reached via the parent of the
hidden field. However, if the field is added again to a different group,
the field element will be moved, thus the parent changes and the
visualization indicator is displayed on the wrong group.
This is a limitation that can only be overcome via condensing or by
enhancing the information in the change.
For now, display removed fields on the form instead to avoid confusion.

SNOW: DINC0056627
Change-Id: I6d8cd0aa19ef326fd57c2f4e21b46b812aab2a47