[FIX] We've introduced the isTrustedSource property as a security enhancement for the PDFViewer control as detailed in the API reference
If isTrustedSource is set to true, the PDFViewer opens with the displayType set to "Embedded" on desktop devices, which means that the PDF content is directly shown within the PDFViewer. Set this property to true only when the PDF is generated by the application or the PDF is scanned for viruses.

If isTrustedSource is set to false, the PDFViewer opens with the displayType set to "Link" on desktop devices, which overrides any configuration that has been provided by the application for the property displayType. This means that the PDFViewer appears as a toolbar with a download button.

PDF's will continue to be displayed by default if the property is not set. However, in future releases, we plan to change this default setting. If not configured manually, PDFViewer will open with displayType set to "Link"

SNOW: CS20240007071117
CR-Id: 002075125900001253752024
Change-Id: Idae71f6dab3b1d36f0c2f5945ee8405e84d60c09