[INTERNAL][FIX] odata.v4.TableDelegate: expand/collapse all rows
Nothing happens if Binding#setAggregation is called with the same value for the expandTo parameter. The expectation is that the state is always reset to this expansion level when pressing the ExpandAll/CollapseAll buttons, also if individual rows were expanded or collapsed since the button was pressed the last time.

Change-Id: I9c9a6c85df2b1ceb471c472b3e40dc05b819f404
SNOW: DINC0061034