[FIX] sap.m.RadioButtonGroup: There is no error when destroying and adding radio buttons on select
After version 1.71 when we have focused element and then we destroy it, it stays focused.
Before 1.71 when we have focused element and then we destroy it, the focused element is the body.
This change causing error in the RadioButtonGroup - after destroying the buttons and adding new buttons (on select)
the group contains focused element (the destroyed button) and tries to focus the selected element,
which is wrong because the new button doesn't have DOM ref yet.
Also if there is no user interaction on the new buttons the focus should not be applied
(selection is done by the application logic).

SNOW: DINC0064975
Change-Id: I59d50f29a3934126b5678bb7abf492a7089fefb9
CR-Id: 002075125800000116022024