[INTERNAL][FIX] MDCTable: Fix type of propertyInfo property
The type object[] is not valid for properties. XML view parsing fails if the propertyInfo property of the table is set and has more than one object in the array. As a result, the app does not even start.
It is recommended to use type object instead.

This is a regression that was introduced in version 1.111 with If34de3014cb16a8b67062772cdb10e10ab16a32d.

Change-Id: Idfd999a8d47f8435fdb8d06e6a5a2bb8ec04ac67
BCP: 2370103929
CR-Id: 002075125900001318342024