[INTERNAL] odata.v4.TableDelegate: Avoid errors on expand/collapseAllRows
odata.v4.ODataListBinding#refresh throws in several cases, for example
if it is not a root binding. The expand/collapseAllRows operation is not
possible in such a case. The faulty solution that was introduced with
I9c9a6c85df2b1ceb471c472b3e40dc05b819f404 is removed entirely.

Change-Id: I1148a7e9f2b26b34637ab60fcfcc57eb1de4fae9
SNOW: DINC0078049