[INTERNAL] MultiInput: creation of tokenizer moved in separate method
The Problem:
As sap.m.MultiInput creates an instance of sap.m.Tokenizer in its init method and sets it to its 'tokenizer' aggregation this leads to an issue when trying to extend and override the tokenizer aggregation because of it's type.

The Solution:
With this change we move creation of tokenizer in a separate method that is private and restricted hence extending the sap.m.MultiInput and sap.m.Tokenizer can be done appropriately.

JIRA: BGSOFUISAKAR-999
Change-Id: I95763ff14e0ce307d5527eb18c0d2da4b8727eb3