[FEATURE] sap.m.Menu, sap.ui.unified.Menu: added MenuItem shortcutText property
* new property shortcutText is introduced in Menu items of sap.m and
sap.ui.unified Menu controls. When set, it displays the shortcut text
in the right side of the Menu item, if there is no submenu of this
item. The shortcut is also set as aria-keyshortcuts attribute of the
item DOM. The shortcut text is not displayed on mobile devices.

JIRA: BGSOFUIBALKAN-7772
Change-Id: Iee7325a6f0cffd2a87e78e683f1cea4af68f5f7a