[INTERNAL] TooltipBase: Temporarily deactivate test
The test currently runs unstable in the voter.
We disable it temporarily to allow for more time to fix it correctly.

The assumption is that a multitude of setTimeout calls (26) leads
to uncaptured asynchronous behavior, most likely related to
the rendering of the TooltipBase.

Change-Id: I8c40c015001c3b9e4e6141d4a074f25ad7b00438
SNOW: n/a
CR-Id: 002075125800000296522024
(cherry picked from commit 25525876537ea8d6c9fb9dc9d6ce739748e64c5d)