[FIX] sap.m.DynamicDateRange: ValueHelp is shown when interacting with keyboard
Problem: ValueHelp was not open when pressing alt + Pg Up or alt + Pg Down

Root cause: showValueHelp is false by default and should be true in order the value help to be opened since DynamicDateRangeInput inherits Input methods onsaphide and onsapshow. However this was leading to under issue on mobile where the value help shouldn't be opened from the input field but only from the icon.
Related to this gerrit change: 6067713

Solution: Overwrite onsaphide and onsapshow methods of the Input where remove the check if the showValueHelp is true or not.

SNOW: DINC0082301
Change-Id: I73a536fee2278a4e71b4071285e6fd67cc749ab4