[FIX] sap.f.DynamicPage: Scrollbar issue fixed
In windows OS (>=10) in Chrome (reported for version 102) there
is a part of the scrollbar (which is wider than in other browsers or
in same browser in Mac) when it runs to the top of the page, overlapped
by the absolute header. Adding ths situation to the specific
selector makes the space for scrollbar wider.

SNOW: CS20240006986712
CR-Id: 002075125800000261552024
Change-Id: I27582d7261d344ab15cc76243d446deab833be9c
(cherry picked from commit 2d50599b7e25d634b9f09a60677c953e2711015b)