[FIX] m.VariantManagement: correction for deletion indicator
Previously the sap.m.VariantItem's property 'visible' was used to
indicate which variant item was deleted in the 'Manage Views' dialog.
This led to issues when a model was bound to the aggregation 'items' of
the VariantManagement control.
This fix uses as the indicator for deleted items an separate, already
maintained array.

BCP: 002075129500001798992024
Change-Id: I39fa2a0171d98de270549ad85900f9c3b866863c
CR-Id: 002075125800000229832024