[FIX] Field: initializing currency updates value
If a currency field is initialized after it had a value. the initial value is directly changed.
Sample:
User enters 1 USD, so value is ["1", "USD"]
After this the field is initialized (via Model), so value is set to [null, ""]
The Field changes to the initial value of the data type ["0", null] what leads to an unwanted update of the model.

Change-Id: Ia14c553be520c813854d4fcb11d78d3b5c264f8c
SNOW: DINC0080072