[INTERNAL][FIX] CellSelector: Adjust column index in GridTable
There are cases, where the column index of the ui.Column does not correctly
represent the current index of the column (is 3, but is actually 2). This
lead to problems in visualization, where the wrong cell is selected.
Instead of retrieving the column index with getIndex(), we simply use the
index from the selection.

Change-Id: Ie53f3d2390ac08da2ca83b0958cffdeb736d9329