[INTERNAL][FIX] CellSelector: Use getElementById instead of closestTo
As IDs often times contain "." or ":" closestTo won't work, as querySelector will not
work with ID's that contain multiple "special" characters.
Therefore, instead of using closestTo to get the element via ID, Element#getElementById will be used.

Additionally, fixed scrolling issue when pressing SPACE and issue with _iBtt flag, which was not initialized
when retroactively adding the CellSelector.

Change-Id: Ib006732e906b99d0ed1efb9d3e8f0bb391478e1c