[INTERNAL] Table: Don't open cell filter menu on summary rows
It is documented that the table's cell filter context menu is not opened when a custom context menu is provided. The cell filter was still opened on summary rows, this is fixed.

Since it is questionable to have the cell filter on summary rows (would filter the table by an aggregated value), it is no longer opened on summary rows in general.

Unused hook features are removed.

Change-Id: I830dff2e3fd53eb7a2d4cf1fe25511c3f1791bc4