[INTERNAL] TypeScript: suppress generation of jQuery from api.json
With the recent change Ia4eb309de91947060c81c343b7ece12aabd41d4a (commit
3c5883545c3391af800293520bb49c4befb99488), jQuery finally became visible
in the api.json as export of module sap/ui/thirdparty/jquery and types
would be generated for it from the JSDoc contained in UI5.

But types for jQuery are already part of the (hard-coded) preamble
and the types generated from the api.json would not be compatible with
those (richer) types.

The generation of the two documented exports of sap/ui/thirdparty/jquery
is therefore suppressed via the .dtsgenrc file.

Change-Id: I99d9771f59427cc534bb44d45386ee396cf3161a