[INTERNAL] sap.ui.fl: Control specific fl delegate registration
This change is an intermediate change to allow the registration of
control specific delegates. The change is not yet fully functional
and will be completed in a follow-up change.

The 'old' delegate functionality is still active and can be used.

Delegate registration is now possible for a specific control. This
allows to register a 'write' delegate for a specific control
additionally to the model specific default 'read' delegate definition.
The read and write delegate registration and getter functions are
all splitted into separate functions.

Change-Id: Ic238a212cfb939be3b3dc5fcd9cbac1699d67229
JIRA: TEAM42-325