[FIX] sap.m.IconTabBar: Adjust responsive paddings for Belize theme
Misalignment has been identified between the paddings
of the IconTabBar, its content and the Header area within the Belize
theme when viewed in desktop mode. To resolve this issue, this change
introduces a modification to the existing paddings specifically for the
Belize theme. This ensures that they're aligned properly across all
components, improving overall UI consistency.

SNOW: CS20240007052684
Change-Id: Icb10f5f80d309f29f0521e6455751020b088e6d1