[INTERNAL] docs/jsdoc.md: Additional example for @deprecated
In case of deprecations, either the user needs information about whether
there is an alternative to the deprecated element, or it needs to be
indicated that the concept has been discarded. This change adds an
additional example for the latter.

It also removes the part about "internal releases" since this isn't
really a common thing anymore and thus doesn't need to be part of these
guidelines.

Also removes the patch level from the version in the example for the#
@experimental tag. Similar to deprecations, it usually should only
mention major and minor version.

Closes https://github.com/SAP/openui5/pull/3999 .

Change-Id: I29b0f8e477118e10be321b29125394624c41b404