[FIX] sap.f.GridContainer: Fix in focus handling for tiles
Solving the following problem:
When use holds down mouse over tile and accidentally slides it outside of a tile,
then releases the mouse, the keyboard navigation inside the grid was not possible anymore.
This was happening because the tile itself is focused instead of the wrapper.

Now this is fixed by handling the mouseup event better.
- listen for mouseup also outside of the grid container
- the focus target is the mousedown target instead of the mouseup target

SNOW: DINC0058793
Change-Id: I7cb60c949f859f9515c30eae8bae66fb9c364eee