[FIX] sap.uxap.ObjectPageLayout: Layout issue corrected
Problem:
Missalignment of the form fields on phone in compact mode.
This happends becauase the ObjectPage overrides the Form
CSS on phone in compact mode(by VD requirement), but does
it only for the case when the form uses a ResponsiveGridLayout
and not for the case when the form uses a ColumnLayout.
So when the page contains two forms, the first with
ResponsiveGridLayout and the second with ColumnLayout,
a missalignment shows.

Solution:
The second case (when the form uses a ColumnLayout)
is analogous to the first, so the same handling
should be applied there as well. Updated the CSS
accordinly.

SNOW: DINC0051118
Change-Id: I5202bf8fea8474fe2225a8066529dcd5dfd56dcf