[FIX] sap.m.Carousel: Resolve issue preventing text selection with mouse
Client reported that after 1.114 highlight and selection of text with mouse was not possible inside Carousel, instead drag was initiated.
Bug was caused due to a missed check for text/editableContent when we refactored the Carousel control.
This change serves to resolve the High priority issue for the client, who is running S/4HANA CE 2402 - currently on 1.120 version.

SNOW: CS20240007032724
Change-Id: I1280d00875ddbb430d3e7ae6cdee876ff10fabc8
CR-Id: 002075125800000738432024