[INTERNAL] Form: change Title size for horizon-theme
To align with other controls the Form-Header Title needs to be rendered in H5 size as default. The rendered HTML level <H4> is not changed to prevent a break in application header levels.

Internal samples changed to use the horizon levels for Titles.

Use horizon theme as default theme für the whole layout library as belize theme is deprecated.

Change-Id: Iea468196d8de789ba1e6fac90f4304cd111e5970