[INTERNAL] sap.ui.unified.Calendar: remove navigation on week selection in two months mode
ISSUE: When selecting a week from the second month, the calendar navigates so the selected month becomes the first one, so navigation is performed. This happens because we hardcoded the "other month" check to be true

SOLUTION: Add a check to verify if the selected interval is part of another month.

SNOW: CS20240007135196

CR-Id: 002075125800002336602024
Change-Id: I0324c47d6ab647776315e5d74804fef6543c957a
(cherry picked from commit ee1f437c13b19fca0fbd7be67cd25b5c9cbdd749)