[INTERNAL] sap.ui.test: Autowaiter now ignores background animations
Problem:
The app may run a background animation [that never ends]. This
causes the jsAutowaiter wait endlessly and the test cannot proceed.

Solution:
Define config parameter 'maxDuration' to the jsAnimationWaiter
(similar to parameter 'maxDelay' of the timeoutWaiter)
and ignore (i.e. do not wait) animations that exceed
the maxDuration value.

SNOW: DINC0093968
Change-Id: I0d1b54a3d26dd6e217ecd07137859eedb0e1dfc6