[INTERNAL] Integration Cards: Provide option to render tile as a link
Now all tiles which have navigation action with a given url will render as link tags.
The behaviour is the same as for GenericTile when given href property.

Added experimental properties href and target to the BaseHeader.
Those properties should be private for now, but then binding will
not work, therefore they are experimental.

JIRA: BGSOFUIRODOPI-3236
Change-Id: I451e4a868d2fa061e3e96e78aad088f19cb54c24