[INTERNAL] sap.m.Avatar: Magnifier icon not shown for fallback image
Problem: When a fallback image is shown (the real one is not found), the
magnifier icon is still shown, if detailBox aggregation is set. However,
it does not make much sense, as there is no image to be enlarged.

Solution: Does not show the magnifier icon, if a fallback image is
shown, no matter if detailBox aggregation is set or not.

JIRA: BGSOFUIPIRIN-6503

Change-Id: Icdd2f5e4b0bd61f527a7117682f17ca1ce3f13f0